refactor: refactoring code to add all the new work from this branch back to can_history_trigger_func
